% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateCaptureEff}
\alias{estimateCaptureEff}
\title{Estimate capture efficiency for simulated data
method to estimate parameter for capture efficiency}
\usage{
estimateCaptureEff(Data, compareData, protocol, fromUMI)
}
\arguments{
\item{Data}{This is the original data from the sce.}

\item{compareData}{This is the initial mRNA counts.}

\item{protocol}{Which scRNA-seq protocol is being simulated.}

\item{sceUMI}{whether or not the original data are UMI counts.}
}
\description{
Estimate capture efficiency for simulated data
method to estimate parameter for capture efficiency
}
