% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateScaffoldParameters}
\alias{estimateScaffoldParameters}
\title{Estimate scaffold parameters from a SingleCellExperiment.}
\usage{
estimateScaffoldParameters(
  sce,
  numCells = NULL,
  numGenes = NULL,
  geneMeans = NULL,
  geneTheta = NULL,
  totalTranscripts = NULL,
  genes = NULL,
  captureEfficiency = NULL,
  efficiencyRT = NULL,
  typeOfAmp = "PCR",
  numFirstAmpCycles = 18,
  numSecondAmpCycles = 12,
  firstAmpEfficiency = NULL,
  secondAmpEfficiency = NULL,
  tagEfficiency = NULL,
  degree = NULL,
  percentRange = -1,
  protocol = "C1",
  totalSD = NULL,
  useUMI = FALSE,
  model = "p",
  usePops = NULL
)
}
\arguments{
\item{sce}{A SingleCellExperiment object used to estimate the scaffold parameters.}

\item{numCells}{The number of cells to be used in the simulation. If left NULL, the number of cells in the \code{sce} object is used.}

\item{numGenes}{The number of genes to be used in the simulation. If left NULL, the number of genes in the \code{sce} object is used.}

\item{geneMeans}{The mean expression level of each gene. If left NULL, the means are estimated using the \code{sce} object.}

\item{geneTheta}{The theta parameter used if the negative binomial distribution is chosen to generate the initial counts for each gene. Should be a numeric vector of length \code{numGenes}. If left NULL, this value is estimated from the \code{sce} object.}

\item{genes}{A vector of names for each gene. If left NULL, the gene names from the \code{sce} object are used.}

\item{captureEfficiency}{A value between 0 and 1 indicating the proportion of mRNA molecules successfully captured for the genes in each cell. If left NULL, this value is estimated using the \code{sce} object.}

\item{typeOfAmp}{The amplification method used in the simulation, defaults to "PCR".}

\item{numFirstAmpCycles}{The number of cycles in the first amplification stage of the simulation.}

\item{numSecondAmpCycles}{The number of cycles in the second amplification stage of the simulation.}

\item{firstAmpEfficiency}{A value between 0 and 1 indicating the efficiency of the first simulated amplification cycle. If set to 1, then all molecules will double each cycle.}

\item{secondAmpEfficiency}{A value between 0 and 1 indicating the efficiency of the second simulated amplification cycle.}

\item{tagEfficiency}{A value between 0 and 1 indicating the tagmentation efficiency.}

\item{degree}{Numeric, determines the amount of perturbation applied to the simulated initial gene counts.}

\item{percentRange}{A value between 0 and 1 indicating the number of samples that undergo exact dilution in the equalization step of the simulation. A negative value indicates that all samples are diluted by the same factor, resulting in no equalization performed.}

\item{protocol}{The protocol to model in the simulation.}

\item{totalSD}{The total sequencing depth of the simulated data. If left NULL, this is taken from the \code{sce} object.}

\item{useUMI}{A TRUE/FALSE indicating whether the protocol should use UMI (Unique Molecular Identifier).}

\item{model}{The probability distribution used to generate the initial gene counts. Defaults to 'p' for the Poisson distribution, but can also be set to 'nb' to use the Negative Binomial distribution.}
}
\description{
Estimate scaffold parameters from a SingleCellExperiment.
}
