% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialCounts.R
\name{generateDynamicGeneCounts}
\alias{generateDynamicGeneCounts}
\title{Generate a dynamic cell population}
\usage{
generateDynamicGeneCounts(numCells, mu, dynamicParams)
}
\arguments{
\item{numCells}{The number of cells to generate counts for.}

\item{mu}{a vector containing the mean of each gene to simulate counts for.}
}
\description{
Generate a dynamic cell population
}
