% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makePlots}
\alias{makePlots}
\title{Generate summary plots of simulated data}
\usage{
makePlots(simulatedData, originalData)
}
\arguments{
\item{simulatedData}{An object of class SingleCellExperiment generated with the \code{simulateScaffold} function.}

\item{originalData}{The original SingleCellExperiment object used to initiate the simulation.}
}
\description{
Generate summary plots of simulated data
}
