% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf_choice.R
\name{scf_choice}
\alias{scf_choice}
\title{scf_choice}
\usage{
scf_choice(choices = character(), name = "item", desc = "choose")
}
\arguments{
\item{choices}{choices t select among or a zero
length prototype e.g. integer(0)}

\item{name}{name to show and to store values in}

\item{desc}{an optional description to show to user}
}
\description{
scf_choice
}
