% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf_copy_files.R
\name{scf_copy_files}
\alias{scf_copy_files}
\title{scf_copy_files}
\usage{
scf_copy_files(..., path = ".")
}
\arguments{
\item{...}{file names to move from}

\item{path}{path to move to}
}
\value{
Nothing, called for side effect on file system.
Copies files from package's inst folder to path.
}
\description{
utility function moving files from inst dir to path
}
