% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf_menu.R
\name{scf_menu}
\alias{scf_menu}
\title{scf_menu}
\usage{
scf_menu(..., .fun = function(...) { })
}
\arguments{
\item{...}{list of choices and inputs}

\item{.fun}{function that makes use of options collected}
}
\value{
nothing, called for its side effects
}
\description{
scf_menu
}
