% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf_scaffold.R
\name{scf_scaffold}
\alias{scf_scaffold}
\title{general scaffolding function}
\usage{
scf_scaffold(scf, path_to = NULL, package = "scaffold",
  overwrite = FALSE)
}
\arguments{
\item{scf}{path to scaffold from}

\item{path_to}{path to scaffold to}

\item{package}{the package to look for the scaffolding}
}
\description{
general scaffolding function
}
\examples{
\dontrun{

scf_scaffold("shiny_material", path_to = ".")

}

}
