% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf_shiny_material.R
\name{scf_shiny_material}
\alias{scf_shiny_material}
\title{scf_shiny_material}
\usage{
scf_shiny_material(path = "./shinyapp")
}
\arguments{
\item{path}{path where to create function}
}
\value{
Nothing, called for side effect on file system.
Creates shinyapp folder needed files for shiny material app.
}
\description{
scf_shiny_material
}
