% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{sequenceStep10X}
\alias{sequenceStep10X}
\title{Sequencing step of the Scaffold simulation for 10X/droplet}
\usage{
sequenceStep10X(
  capturedMolecules,
  totalDepth,
  efficiencyPCR,
  roundsPCR,
  efficiencyTag,
  genes,
  useUMI
)
}
\description{
Sequencing step of the Scaffold simulation for 10X/droplet
}
