% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{sequenceStepC1}
\alias{sequenceStepC1}
\title{Sequencing step of the Scaffold simulation}
\usage{
sequenceStepC1(
  amplifiedMolecules,
  equalizationAmount,
  totalDepth,
  efficiencyPCR,
  roundsPCR,
  efficiencyTag,
  genes,
  useUMI
)
}
\description{
Sequencing step of the Scaffold simulation
}
