% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateScaffold}
\alias{simulateScaffold}
\title{Run a scaffold simulation.}
\usage{
simulateScaffold(
  scaffoldParams,
  originalSCE,
  model = "p",
  inputInitial = NULL,
  SD
)
}
\description{
Run a scaffold simulation.
}
