% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uneq_ec_data}
\alias{uneq_ec_data}
\title{An example, unprocessed single-cell data set}
\format{
A numeric matrix of gene counts.
}
\usage{
uneq_ec_data
}
\description{
An example, unprocessed single-cell data set
}
\keyword{datasets}
