% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{captureStep}
\alias{captureStep}
\title{Capture cDNA from cell (cell lysis and conversion of mRNA to cDNA)}
\usage{
captureStep(
  Data,
  captureEffCell = NULL,
  captureEffGene = NULL,
  rtEffCell = NULL,
  rtEffGene = NULL,
  useUMI = FALSE
)
}
\description{
Capture cDNA from cell (cell lysis and conversion of mRNA to cDNA)
}
