% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateScaffoldParameters}
\alias{estimateScaffoldParameters}
\title{Estimate scaffold parameters from a SingleCellExperiment.}
\usage{
estimateScaffoldParameters(
  sce,
  numCells = NULL,
  numGenes = NULL,
  geneMeans = NULL,
  geneTheta = NULL,
  genes = NULL,
  captureEfficiency = NULL,
  typeOfAmp = "PCR",
  numFirstAmpCycles = 18,
  numSecondAmpCycles = 12,
  firstAmpEfficiency = NULL,
  secondAmpEfficiency = NULL,
  tagEfficiency = NULL,
  degree = 5,
  percentRange = -1,
  protocol = "C1",
  totalSD = NULL
)
}
\description{
Estimate scaffold parameters from a SingleCellExperiment.
}
