% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialCounts.R
\name{generateGeneCounts}
\alias{generateGeneCounts}
\title{Generate initial counts}
\usage{
generateGeneCounts(numCells, mu, popHet)
}
\arguments{
\item{numCells}{The number of cells to generate counts for.}

\item{mu}{a vector containing the mean of each gene to simulate counts for.}

\item{popHet}{the degree of heterogeneity within the cell population.
Should be a vector of length two specifying the bounds.}
}
\value{
Output is matrix of counts, genes on the rows and cells
  on the columns.
}
\description{
Generates the transcripts in the cell.
}
