% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplification.R
\name{preamplifyStep}
\alias{preamplifyStep}
\title{Pre-amplification step}
\usage{
preamplifyStep(
  capturedMolecules,
  genes,
  efficiencyPCR,
  rounds,
  typeAMP,
  useUMI
)
}
\description{
Pre-amplification step
}
